move bytes_to_field, etc from conversion.rs to new bytes_to_field.rs
